separate the image and volume classes into different folders
