+++ src/aspyre/image/image.py
