+++ tests/test_covar.py
@@ -9,7 +9,7 @@ import pytest
 from aspyre.source import SourceFilter
 from aspyre.source.simulation import Simulation
 from aspyre.basis.fb_3d import FBBasis3D
-from aspyre.imaging.filters import RadialCTFFilter
+from aspyre.utils.filters import RadialCTFFilter
 from aspyre.estimation.mean import MeanEstimator
 from aspyre.estimation.covar import CovarianceEstimator
 from aspyre.utils.matrix import eigs
