+++ tests/test_filters.py
@@ -2,7 +2,7 @@ import numpy as np
 from unittest import TestCase
 
 from aspyre.source import SourceFilter
-from aspyre.imaging.filters import RadialCTFFilter
+from aspyre.utils.filters import RadialCTFFilter
 
 import os.path
 DATA_DIR = os.path.join(os.path.dirname(__file__), 'saved_test_data')
