+++ tests/test_white_noise.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 from aspyre.source import SourceFilter
 from aspyre.source.simulation import Simulation
-from aspyre.imaging.filters import RadialCTFFilter
+from aspyre.utils.filters import RadialCTFFilter
 from aspyre.estimation.noise import WhiteNoiseEstimator
 
 import os.path
